requirements.txt
# Trabalho final IN242
#Prof.: Lúcio

paho-mqtt
pymongo